Refactor: 함수 위치 이동
